+++ src/com/qmetric/model/pricingmodels/CostPricingModel.java
@@ -1,5 +1,6 @@
 package com.qmetric.model.pricingmodels;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 
 import com.qmetric.actors.Supplier;
@@ -11,7 +12,7 @@ public class CostPricingModel implements PricingModel {
 	private Supplier supplier;
 	private StockUnitsOfMeasure units;
 	private Currency curr;
-	private int price;
+	private BigDecimal price;
 	private Calendar dateInForce;
 
 	@Override
@@ -25,7 +26,7 @@ public class CostPricingModel implements PricingModel {
 	}
 
 	@Override
-	public int getPriceInCents() {
+	public BigDecimal getPriceInCents() {
 		return price;
 	}
